Removed some compiler warnings / added missing initializations

+++ multitude/Nimble/Math.hpp
@@ -276,14 +276,14 @@ namespace Nimble {
         inline void calculateMeanVariance(const T * values, int n, T * mean, T * variance)
     {
       T ave = 0;
-      for(size_t i = 0; i < n; i++)
+      for(int i = 0; i < n; i++)
         ave += values[i];
 
       ave /= (double) n;
       *mean = ave;
 
       T vari = 0;
-      for(size_t i = 0; i < n; i++) {
+      for(int i = 0; i < n; i++) {
         T tmp = values[i] - ave;
         vari += tmp * tmp;
       }
@@ -298,7 +298,7 @@ namespace Nimble {
     inline T calculateAbsSum(const T * values, int n)
     {
       T sum = 0;
-      for(size_t i = 0; i < n; i++)
+      for(int i = 0; i < n; i++)
         sum += std::abs(values[i]);
 
       return sum;
